[src] Add scale factor in vibesMaze

+++ examples/cpp/invariants/2D/22_van_der_pol_invariant_bracketing_stats/main.cpp
@@ -69,7 +69,7 @@ int main(int argc, char *argv[])
         // ******* Algorithm ********* //
         double time_start = omp_get_wtime();
 
-        for(int i=0; i<20; i++){
+        for(size_t i=0; i<20; i++){
             cout << i << endl;
             double time_local = omp_get_wtime();
